fabsf

+++ include/common.h
@@ -14,5 +14,5 @@ typedef float f32;
 
 #define NULL 0
 
-double fabs(double f);
-#pragma intrinsic(fabs)
+double fabsf(double f);
+#pragma intrinsic(fabsf)
